[ 004 ] Library logging in VERBOSE mode
My mistake lead to unwanted behaviour of the logging system. It's now fixed by making the library log the information to the stdout/err ONLY IF the build system defines the 'VERBOSE' macro. That way we can keep the logging functions (gf_logx) for the user AND for the library without limiting it to the build mode. This is also a moment when I should consider making a proper Makefile (and build systems in general).