[ 007 ] More window events
I've added move, mouse motion, mouse button press and release events to gf. Also, in the header file there's now a little documentation of how all the events utilize the data section of the event. With addition of move event, there's now a new API function: gf_getWindowPosition. With this addition x11 backend not only tracks the width and height of the window, but it's position as well.