[ 030 ] Further modularization and config
Internal modularization proceeds with abstracting client and context to their respective structs, defined in the internal headers. Internal includes were also moved to internal headers. This will help in the future with dynamic loading of the APIs. Going further, I've started to finally implement a proper config system. For now it only supports resizable and maybe borderless (can't be sure on i3wm).