[ 032 ] X11: More configuration options
gf now supports more configuration options, implemented only for X11 as of now. Wayland module was completely eradicated because I want to work on it better in the future when I'll be happy with the current state of Xlib support. As of now I still need to test the configuration options on other DE's, make a whole dynamic loading module and polish the codebase.