[ 035 ] EGL: Dynamic library loading
In the same fashion as X11, EGL is now being loaded dynamically during the runtime of the prorgam, which means we don't need to link with '-lEGL'. Also, the same as X11, we're only loading currently used programs. In the future, after I finish writing the loader for GLX, I'll try to get most of the functions loaded by gf.